allow filesystems to start with /dpm.*/ for http
changed the template for zlcgdm-dav.conf to send all requests
to /<ns_prefix>/ and not /<ns_prefix> to the nameserver and
the rest to the disk node part. In effect, you now can have
a filesystem starting with /dpm (not /dpm/, but e.g. /dpmfs).
Also, it only applies to the apache frontend and restrictions
for other protocols may apply.