LCGDM-1503: add -k fifo to xrootd options for xrootd4
Add the parameter '-k fifo' hardcoded to the xrootd sysconfig _OPTIONS
lines. Do not yet remove any .*.log files yet, but try if it works
without doing that.

This patch does not make any attempt to find out the xrootd version (3.x
or 4.x) so it should be only used with xrootd4.