LCGDM-1503: deleting unwanted .xrootd.log files
only changing the config parameter for "-k fifo" does not work, because
xrootd fails if the .xrootd.log files exist, but are regular files.
Delete all of them if "-k fifo" is set and there are any regular files,
or if "-k fifo" is not set and there are any pipes.
This should reduce the number of deletions as much as possible.
Xrootd is restarted after deleting any files.