add simple configuration for the profiler monitoring
there are no configurable except the collectors, since the monitoring
only workds with one config. the collectors have to be passed as a list
of strings (this is also validated in the code).
we execpt that both /etc/dmlite.conf.d/ and /etc/dmlite-disk.conf.d/
exist, which both head.pp and disk.pp provide.